Fix comment attachment retries and share document model
